refactor: moved exceptions into separated files
