+++ src/main/java/com/vodafone/exception/cart/NegativeQuantityException.java
@@ -1,4 +1,4 @@
-package com.vodafone.exception;
+package com.vodafone.exception.cart;
 
 public class NegativeQuantityException extends RuntimeException {
     public NegativeQuantityException(String message) {
