+++ src/main/java/com/vodafone/exception/cart/NullCartItemException.java
@@ -1,4 +1,4 @@
-package com.vodafone.exception;
+package com.vodafone.exception.cart;
 
 public class NullCartItemException extends RuntimeException{
     public NullCartItemException(String message){
