+++ src/main/java/com/vodafone/exception/customer/IncompleteUserAttributesException.java
@@ -1,4 +1,4 @@
-package com.vodafone.exception;
+package com.vodafone.exception.customer;
 
 public class IncompleteUserAttributesException extends RuntimeException{
     public IncompleteUserAttributesException(String message){
