+++ src/main/java/com/vodafone/exception/customer/NullCustomerException.java
@@ -1,4 +1,4 @@
-package com.vodafone.exception;
+package com.vodafone.exception.customer;
 
 public class NullCustomerException extends RuntimeException{
     public NullCustomerException(String message){
