+++ src/main/java/com/vodafone/service/CustomerService.java
@@ -1,10 +1,9 @@
 package com.vodafone.service;
 
-import com.vodafone.exception.IncompleteUserAttributesException;
-import com.vodafone.exception.NullCustomerException;
+import com.vodafone.exception.customer.IncompleteUserAttributesException;
+import com.vodafone.exception.customer.NullCustomerException;
 import com.vodafone.exception.NullIdException;
 import com.vodafone.model.Customer;
-import com.vodafone.repository.customer.CustomerRepository;
 import com.vodafone.repository.customer.ICustomerRepository;
 import org.springframework.stereotype.Service;
